scripts/requirements.txt
fake==8.12.1
requests